[refactoring] CommandFactory h/cpp 분리
