+++ test-shell/command_factory.h
@@ -1,10 +1,13 @@
 #pragma once
+
 #ifndef interface
 #define interface struct
 #endif
+
 #include <string>
 #include <memory>
 #include <unordered_map>
+#include <iostream>
 #include "command.h"
 
 interface ICommandFactory {
@@ -14,4 +17,27 @@ interface ICommandFactory {
 	virtual const std::unordered_map<std::string, ICommand*>&
 		getAllCommands() const = 0;
 	virtual void initialize(ISSD* ssd) = 0;
+};
+
+
+class CommandFactory : public ICommandFactory {
+public:
+	static ICommandFactory& getInstance();
+
+	void initialize(ISSD* ssd) override;
+
+	void injectCommand(const std::string& name,
+		ICommand* command) override;
+
+	ICommand* getCommand(const string& command) override;
+
+	const std::unordered_map<std::string, ICommand*>& getAllCommands() const override;
+
+private:
+	CommandFactory();
+
+	CommandFactory(const CommandFactory&) = delete;
+	CommandFactory& operator=(const CommandFactory&) = delete;
+
+	std::unordered_map<std::string, ICommand*> commands;
 };
