+++ unit-test/test_main.cpp
@@ -9,3 +9,5 @@
 #include "../test-shell/TestApp2Command.cpp"
 #include "../test-shell/EraseRangeCommand.cpp"
 #include "../test-shell/EraseCommand.cpp"
+#include "../test-shell/command_factory.cpp"
+#include "../test-shell/shell.cpp"
